codeschool es6 - level-1 - Declarations with let
Understanding Hoisting
Prior to executing our code, JavaScript moves var declarations all the way up
to the top of the scope. This is known as hoisting.

let variables are scoped to the nearest block and are NOT hoisted.
(A block is any code section within curly braces, like if, else, for, while, etc.)

Using let, variables are “trapped” inside their respective if and else blocks.