codeschool es6 - level-1 - Declarations with let part 2 - Declarations With let in for Loops
Understanding Hoisting and for Loops
var i is hoisted to the top of the function and shared across each iteration of the loop.

Loop Values in Callbacks
When callbacks begin to run, i holds the last value assigned to it from the for loop.

Using let in for Loops
With let, there’s no sharing in for loops. A new variable is created on each iteration.

Using let in for Loops
Each callback function now holds a reference to their own version of i.

let Cannot Be Redeclared