codeschool es6 - level-4-1 - Array
Assigning From Array to Local Variables
Reading Values With Array Destructuring
We can use destructuring to assign multiple values from an array to local variables.

Using for...of to Loop Over Arrays
The for...of statement iterates over property values, and it’s a better way to loop over arrays and other iterable objects.

Objects That Don’t Work With for...of
No function assigned to the Symbol.iterator property means the object is not iterable.

Finding an Element in an Array
Array.find returns the first element in the array that satisfies a provided testing function.