codeschool es6 - level-4-2 - Map
Storing Key/Values With Map
The Map object is a simple key/value data structure. Any value may be used as either a key or a value, and objects are not converted to strings.

Iterating Maps With for...of
Maps are iterable, so they can be used in a for…of loop. Each run of the loop returns a [key, value] pair for an entry in the Map.

WeakMap
The WeakMap is a type of Map where only objects can be passed as keys. Primitive data types — such as strings, numbers, booleans, etc. — are not allowed.

WeakMaps Are Better With Memory
Individual entries in a WeakMap can be garbage collected while the WeakMap itself still exists.