codeschool es6 - level-4-3 - Set
Limitations With Arrays
Arrays don’t enforce uniqueness of items. Duplicate entries are allowed.

Using Set
The Set object stores unique values of any type, whether primitive values or object references.

Using Set as Enumerable Object
Set objects are iterable, which means they can be used with for...of and destructuring.

WeakSet
The WeakSet is a type of Set where only objects are allowed to be stored.

Can’t Read From a WeakSet
WeakSets cannot be used with for...of and they offer no methods for reading values from it.